tree: return types ilPathGUI::getPath();

+++ Services/Tracking/classes/repository_statistics/class.ilTrSummaryTableGUI.php
@@ -720,7 +720,7 @@ class ilTrSummaryTableGUI extends ilLPTableBaseGUI
             // #16453
             include_once './Services/Tree/classes/class.ilPathGUI.php';
             $path = new ilPathGUI();
-            $path = $path->getPath($this->ref_id, array_pop($a_set["ref_ids"]));
+            $path = $path->getPath($this->ref_id, (int) array_pop($a_set["ref_ids"]));
             if ($path) {
                 $this->tpl->setVariable('COLL_PATH', $this->lng->txt('path') . ': ' . $path);
             }
